mati/requirements.txt
Flask>=2.2
Flask-Cors>=3.0.10
python-dotenv>=1.0.0
groq>=0.1.0
gTTS>=2.3.0
pytesseract>=0.3.10
Pillow>=9.4.0
requests>=2.28.2